Fixes Issue 397
Final fixes to finish off this issue.

Altered to code so that the updated on/by (metadata) information on a table is only updated when data in that table is actually altered. The exception to this are sample/occurrence attributes which also update the parent occurrence/sample metadata.

e.g. Altering an occurrence no longer results in the parent sample metadata getting updated.