New version supports the following,
Supports Normal and Admin users for the Cudi-Form for viewing boundaries list and adding boundaries

Supports the system assuming that if no preferred boundary is set, then the latest item is the preferred one.

Reloads the cudi-form to show boundary as soon as one is selected.

Supports an update existing boundary option checkbox.

Has a Set Preferred And Save Now button on the boundary list.

Has a Set Preferred On Save checkbox as well, near the submit button on the second tab, if the user wishes to make the item preferred as soon as they save (for instance if they are creating a new one, although it also works with saving existing)

Supports saving when viewing a either count unit or a boundary using the above options (adjusting the boundary when viewing a count unit adjusts the preferred boundary).


Note: That filling in attributes and fields other than the name, centroid and boundary is currently untested. This may or may not work