Further bug fixes and full support for displaying an extra layer showing the feature the user just clicked on. This now works correctly with the page breadcrumb and location drop-down.
Also changed the code, so on layers where Count Units are to be displayed as well as the other location types, clicking on a count unit will go to the Count Unit Information Sheet instead of attempting to navigate to the layer below.

Changed the php file to supply the count unit boundary id as indiciaData.

Also fixed a problem with the code (which did not affect the program at the time) where the $options variable which holds options from the edit tab was being overwritten with map options which in turn was being overwritten by report options.
These different options now all use different variable names.