Working version of the milestones module.
Note there are a few known minor issues with it which will be cleaned up when I get a chance, but do not affect the general working of the module

1. If validation fails then filter settings need to be re-entered
2. Filter created_by_id is being set to the updated_by_id when just updating the filter on existing milestones
3. Text is wrong on the first filter tab
4. Sources tab on the filter not operating in edit mode
5. Am sure there is some room for speeding up the code
6. The setting I put in place so that only certain websites need to be checked for milestones is ignored, all websites are checked



There is also one limitation that may not be considered a problem depending on how you look at it, therefore this behaviour will only be altered if requested as it is not in the specification either,
If milestone changes, then milestone_awards are not removed, a person will still be considered to have passed an award even if the count changes and they haven't reached that count yet.
For this reason, a new milestone should be used when increasing the award count, but can be edited when reducing the count.