Commit 3 of 3
Custom attributes with dates now load with the format specified by a variable in helper_base (no time at the moment to make a proper option).
Sample dates also changed to load with this format.

I have made a note by the variable in helper_base to say that there is probably further room for improvement as am sure there are areas which should but won't respect the date format elsewhere.