New form for Seasearch project. Uses 3 level sample structure. Main sample Subsample (habitats) Third level samples (holds spatial references for occurrences) Occurrences
The form is progressive in nature, as the user moves through the wizard, the data structure matures and becomes more complex. So at the start, the user only uploads photos to a main sample.
Only later in the wizard do the other layers come into play.
This means the form needs to use altered versions of the methods found in the normal submission builder and data entry helper, so it is quite a bit form as it includes this code which is normally accessed outside the form